Removal NumberInterface, Money class to UnitTestClass, removal init() methods in AbstractNumber implementations, moved static create() to AbstractNumber, renamed formatCurrency => formatMoney, general improvements
