+++ src/Number.php
@@ -8,19 +8,6 @@ use Number\Formatter\Formatter;
 
 class Number extends AbstractNumber
 {
-    /**
-     * @param string|float|int $value
-     */
-    public static function create($value): self
-    {
-        return new self($value);
-    }
-
-    public function init(string $value): self
-    {
-        return new self($value, $this);
-    }
-
     public function format(int $minFractionDigits = 0, int $maxFractionDigits = 2): string
     {
         return Formatter::format($this->get(), $minFractionDigits, $maxFractionDigits);
